chore: reorganize files
move all server source code into `src/` directory, move the `postprocess` file into `bin/`
