+++ .github/workflows/flat.yml
@@ -28,5 +28,5 @@ jobs:
         uses: githubocto/flat@v3
         with:
           http_url: https://www.sos.state.mn.us/media/3338/precinct-table.xlsx
-          downloaded_filename: lib/data/precinct-table.xlsx # The http_url gets saved and renamed in our repository as this file
-          postprocess: lib/postprocess.js # A postprocessing javascript or typescript file written in Deno
+          downloaded_filename: src/lib/data/precinct-table.xlsx # The http_url gets saved and renamed in our repository as this file
+          postprocess: bin/postprocess.mjs # A postprocessing javascript or typescript file written in Deno
