+++ .prettierignore
@@ -1,2 +1,2 @@
-lib/data
+src/lib/data
 *.hbs
