+++ __test__/build-maps-url.test.js
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 
-const buildMapsUrl = require('../lib/build-maps-url');
+const buildMapsUrl = require('../src/lib/build-maps-url');
 
 describe('#buildMapsUrl', () => {
   it('should return URL with place ID if it is passed', () => {
