+++ __test__/geocode-address.test.js
@@ -2,8 +2,8 @@ import { describe, it, expect } from 'vitest';
 
 const nock = require('nock');
 
-const { GeocodingError } = require('../lib/errors');
-const { forwardGeocode, reverseGeocode } = require('../lib/geocode-address');
+const { GeocodingError } = require('../src/lib/errors');
+const { forwardGeocode, reverseGeocode } = require('../src/lib/geocode-address');
 
 describe('#forwardGeocode', () => {
   it('should return valid data for valid query', async () => {
