--- lib/shorten-address.js
@@ -1,3 +1,0 @@
-module.exports = function shortenAddress(address) {
-  return address.replace(/, USA$/, '');
-};
