+++ src/lib/shorten-address.js
@@ -1,0 +1,3 @@
+module.exports = function shortenAddress(address) {
+  return address.replace(/, USA$/, '');
+};
